test function that returns collection stats
