+++ tests/conftest.py
@@ -8,7 +8,6 @@ import pymongo
 import pytest
 import yaml
 from cyvcf2 import VCF
-from flask_login import login_user, logout_user
 
 # Adapter stuff
 from mongomock import MongoClient
@@ -68,6 +67,7 @@ from scout.parse.panel import parse_gene_panel
 from scout.parse.variant import parse_variant
 from scout.parse.variant.headers import parse_rank_results_header
 from scout.server.blueprints.login.models import LoginUser
+from scout.server.app import create_app
 from scout.utils.handle import get_file_handle
 from scout.utils.link import link_genes
 
@@ -77,6 +77,22 @@ REAL_DATABASE = "realtestdb"
 # root_logger = logging.getLogger()
 # init_log(root_logger, loglevel='INFO')
 LOG = logging.getLogger(__name__)
+
+
+@pytest.fixture
+def mock_app(real_populated_database):
+    """Return the path to a mocked app object with data"""
+    _mock_app = create_app(
+        config=dict(
+            TESTING=True,
+            DEBUG=True,
+            MONGO_DBNAME=REAL_DATABASE,
+            DEBUG_TB_ENABLED=False,
+            LOGIN_DISABLED=True,
+        )
+    )
+    return _mock_app
+
 
 ##################### Gene fixtures #####################
